requirements.txt
langgraph>=0.3.0
langchain>=0.2.0
langchain-openai>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
jinja2>=3.1.0
mysql-connector-python>=8.0.0
